requirements.txt
# 核心依赖
requests>=2.28.0
pyyaml>=6.0
loguru>=0.6.0

# 昇思MindSpore框架（可选）
# mindspore>=2.0.0

# 图像处理（可选）
# Pillow>=9.0.0
# opencv-python>=4.5.0

# API服务
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.28.0

# gRPC（可选）
grpcio>=1.50.0
grpcio-tools>=1.50.0

# 数据库
pymysql>=1.0.2
redis>=4.3.0

# 配置文件
pyyaml>=6.0
python-dotenv>=0.20.0

# 工具类
loguru>=0.6.0
pandas>=1.4.0
scikit-learn>=1.1.0

# AI智能体框架
crewai>=0.70.0
crewai-tools>=0.12.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# 测试框架
pytest>=7.4.0
pytest-cov>=4.1.0
